update to the Qn calculation
I have updated the compute_Qn function to take more than one glass former. I have tested it now for a borosilicate glass that I had previously and the results are the same as the ones from ovito.
The function now dumps the partial Qn as well if many glass formers exist.