high temperature viscosity workflow
This is a first attempt to produce viscosity from MD simulations using the Green-Kubo method.

Everything used here is only for testing purposes and by no means represents an expected value of the viscosity of the used sample for many reasons (i.e., small system size, short simulation time, fewer averages, etc.)

The scripts generate a structure and equilibrate its density and energy, and after that, a production run is used to get the pressure tensor for viscosity calculation.

I plan to move the post-processing step, currently written in the Jupyter notebook, to a function that takes the simulation output as input and returns the viscosity as output.