new simulation  (#56)
Added new single MD simulation at constant T/P that uses the output of another simulation. This comes handy to separate the equilibration step from the production step for the property calculation, so for the calculation of viscosity, CTE, or elastic constants we feed only equilibrated structure and perform production runs directly.

Changed the structure analysis tools to take the ASE Atoms object directly and an input.

Few optimizations of some functions here and there.